requirements.txt
flask==3.0.2
hume==0.7.8
python-dotenv==1.0.1 